TESTS: consolidate polars Surv.from_dataframe tests into test_util
The pandas and polars case factories share one driver by concatenating
their parametrize case lists; polars-specific cases and the cross-backend
error-message test are kept as dedicated tests. Collected test count (35)
and coverage are unchanged.
